Added Pushover Notification to kill_more_than.py
For those who use Pushover (like me) to get notified when things happen with their Plex server, I added a section for users to add a notification when the script fires. 

I added it commented, as I'm not sure the best way to make this optional aside from just removing the comment hashtag.