Update gmusic_playlists_to_plex.py
Occasionally, plex music index returns NoneType and causes the script to error prematurely. We should add a default value to the int conversion to prevent this from happening with int(pmusic.index or 0).

Error:
TypeError: int() argument must be a string, a bytes-like object or a number, not 'NoneType'

Outcome:
By simply forcing 0 for the track number, int(pmusic.index) the comparison between Plex and GPM fails, leading the script to fall back to a track duration match or title match instead of erroring out.